requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
django-environ==0.11.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
